requirements.txt
requests==2.21.0
psycopg2==2.8
python-escpos==2.2.0
SQLAlchemy==1.3.2